python/requirements.txt
grpcio==1.0.0
ipython==5.1.0
